Protecting Videos Routes

+++ controllers/videoController.js
@@ -33,7 +33,7 @@ export const postUpload = async (req,res)=> {
     });
     req.user.videos.push(newVideo.id);
     req.user.save()
-    console.log(routes.videoDetail(newVideo.id)); 
+    // console.log(routes.videoDetail(newVideo.id)); 
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -43,7 +43,11 @@ export const getEditVideo = async (req, res)=> {
         }=req;
     try {
         const video = await Video.findById(id)
-        res.render("editVideo",{pageTitle: `Edit ${video.title}`,video})
+        if( `${video.creator}` !== `${req.user.id}`){
+            throw Error();
+        } else{
+            res.render("editVideo",{pageTitle: `Edit ${video.title}`,video})
+        }
     } catch(error){
         console.log(error);
         res.redirect(routes.home)
@@ -69,7 +73,13 @@ export const deleteVideo = async(req, res)=> {
         params: {id}
     } = req;
     try {
-        await Video.findOneAndDelete({_id:id});
+        const video = await Video.findById(id);
+        if(`${video.creator}` !== `${req.user.id}`){
+            throw Error();
+        } 
+        else {
+            await Video.findOneAndDelete({_id:id});
+        }
     }
     catch(error){
         console.log(error);
@@ -81,7 +91,7 @@ export const videoDetail = async(req, res)=> {
     const {
         params:{id}
     } = req;
-    console.log(req.params);
+    // console.log(req.params);
     try{
         const video = await Video.findById(id).populate("creator");
         console.log(video);
